TG-9: add postman collections and user validations and change status of controller

+++ src/controllers/super-admins.js
@@ -1,24 +1,49 @@
 import superAdmins from '../models/Super-admins';
 
+const { ObjectId } = require('mongoose').Types;
+
+const isValidObjectId = (id) => {
+  if (ObjectId.isValid(id)) {
+    if ((String)(new ObjectId(id)) === id) { return true; }
+    return false;
+  }
+  return false;
+};
+
 const deleteSuperAdmins = async (req, res) => {
+  if (!isValidObjectId(req.params.id)) {
+    return res.status(400).json({
+      message: `Invalid id: ${req.params.id}`,
+      error: true,
+    });
+  }
   try {
     const { id } = req.params;
     const result = await superAdmins.findByIdAndDelete(id);
-    return res.status(200).json({
-      message: 'Super admin deleted',
-      data: result,
+    if (!result) {
+      return res.status(404).json({
+        message: `Super admin not found with id: ${id}`,
+        error: true,
+      });
+    }
+    return res.status(204).json({
       error: false,
     });
   } catch (error) {
-    return res.status(404).json({
-      message: 'Id non-existent',
-      data: undefined,
+    return res.status(500).json({
+      message: `Unexpected error ${error}`,
       error: true,
     });
   }
 };
 
 const editSuperAdmins = async (req, res) => {
+  if (!isValidObjectId(req.params.id)) {
+    return res.status(400).json({
+      message: `Invalid id: ${req.params.id}`,
+      error: true,
+    });
+  }
   try {
     const { id } = req.params;
     const result = await superAdmins.findByIdAndUpdate(
@@ -26,16 +51,14 @@ const editSuperAdmins = async (req, res) => {
       { ...req.body },
       { new: true },
     );
-
     return res.status(200).json({
-      message: 'Super admin edited',
+      message: `Modified Super admin with id ${id}`,
       data: result,
       error: false,
     });
   } catch (error) {
-    return res.status(400).json({
-      message: 'Super admin cannot be edited',
-      data: undefined,
+    return res.status(500).json({
+      message: `Unexpected error ${error}`,
       error: true,
     });
   }
